worker/requirements.txt
pydexcom
python-dotenv
sqlalchemy
psycopg[binary]